+++ turbo/apps/host-worker/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [1.5.61](https://github.com/okou-ai/okou/compare/host-worker-v1.5.60...host-worker-v1.5.61) (2026-09-26)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @okouai/api-contracts bumped to 1.514.3
+
+## [1.5.60](https://github.com/okou-ai/okou/compare/host-worker-v1.5.59...host-worker-v1.5.60) (2026-09-26)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @okouai/api-contracts bumped to 1.514.2
+
 ## [1.5.59](https://github.com/okou-ai/okou/compare/host-worker-v1.5.58...host-worker-v1.5.59) (2026-09-26)